refactor: DTO 객체의 패키지 경로를 더 적합한 곳으로 변경
