--- src/main/java/blackjack/dto/ResultDto.java
@@ -1,7 +1,0 @@
-package blackjack.dto;
-
-import blackjack.model.BlackjackResult;
-
-public record ResultDto(String playerName, BlackjackResult result) {
-
-}
