+++ src/main/java/blackjack/view/dto/CardDto.java
@@ -1,4 +1,4 @@
-package blackjack.dto;
+package blackjack.view.dto;
 
 import blackjack.model.Card;
 import blackjack.model.Rank;
