+++ src/main/java/blackjack/view/dto/ResultDto.java
@@ -1,0 +1,9 @@
+package blackjack.view.dto;
+
+import blackjack.model.BlackjackResult;
+
+public record ResultDto(
+        String playerName,
+        BlackjackResult result
+) {
+}
